Move Terminal and AccessibilityManager into browser
